+++ frontend/src/components/table-layout/TableLayout.vue
@@ -1,26 +1,19 @@
 <template>
   <DefaultLayout>
-    <template #dream-mall-button="{ close }">
-      <TableSettings ref="tableSettingsRef" @close="close" />
-    </template>
-    <template #default>
-      <div class="container">
-        <EmbeddedTable v-if="errorMessage === null" :url="tableUrl" @table-closed="onTableClosed" />
-        <div v-else class="test-not-found">
-          {{ $t('table.notFound') }}
-        </div>
+    <div class="container">
+      <EmbeddedTable v-if="errorMessage === null" :url="tableUrl" @table-closed="onTableClosed" />
+      <div v-else class="test-not-found">
+        {{ $t('table.notFound') }}
       </div>
-    </template>
+    </div>
   </DefaultLayout>
 </template>
 
 <script setup lang="ts">
 import { navigate } from 'vike/client/router'
-import { ref } from 'vue'
 
 import useDreamMallPanel from '#components/dream-mall-panel/useDreamMallPanel'
 import EmbeddedTable from '#components/embedded-table/EmbeddedTable.vue'
-import TableSettings from '#components/malltalk/settings/TableSettings.vue'
 import DefaultLayout from '#layouts/DefaultLayout.vue'
 
 defineProps<{
@@ -28,11 +21,9 @@ defineProps<{
   errorMessage: string | null
 }>()
 
-const { setComponentRef } = useDreamMallPanel()
+const { setMode } = useDreamMallPanel()
 
-const tableSettingsRef = ref<InstanceType<typeof TableSettings>>()
-
-setComponentRef(tableSettingsRef, 'table-settings')
+setMode('table-settings')
 
 const onTableClosed = () => navigate('/')
 </script>
